Shuffle things around a little
Change some layout to better match the standard
as when generated with `bundle gem <new_gem_name>`.
This means the pupistry executable is now in `exe`,
not `bin`.

Also, add the barest of testing frameworks, just
a minitest helper and a failing example test.
Also set some dev dependencies for all this.