Added support for legacy CentOS 5 and CentOS 6 OS releases.
Whilst not exactly the newest and shiniest releases, these versions
still exist in a lot of locations, and by having support, it makes
transistion existing systems to a fully Puppetised environment a lot
easier.

Note that CentOS 5 is the earliest we'll ever go, before then Puppetlabs
doesn't release packages and dependencies and the demand is too low.